Added Edit phase tests
