+++ src/test/java/project/app/warzone/Features/EditPhaseTest.java
@@ -1,30 +1,27 @@
 package project.app.warzone.Features;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.util.LinkedHashMap;
-import java.util.List;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import project.app.warzone.Model.Continent;
 import project.app.warzone.Model.GameEngine;
 import project.app.warzone.Model.Map;
-import project.app.warzone.Model.Node;
 import project.app.warzone.Model.Phase;
 import project.app.warzone.Utilities.MapResources;
 
-public class PreloadTest {
+public class EditPhaseTest {
 
 
     MapFeatures mapFeatures = MapFeatures.getInstance();
     public MapResources mapResources;
     public GameEngine gameEngine;
+    public Phase playPhase;
     
     @BeforeEach
     public void setUp(){
@@ -32,27 +29,30 @@ public class PreloadTest {
         this.mapFeatures = new MapFeatures(mapResources);
     }
 
+    //checks if game is not skippping phase directly to next phase corretly in preloading
+    @Test
+    public void inEditPrePhaseValidate(){
+        String l_mapFileName = System.getProperty("user.dir") + "/src/main/java/project/app/warzone/Utilities/Maps/" + "europe.map";
+        Map l_map = new Map();
+        l_map=mapFeatures.readMap(l_mapFileName);
+        GameEngine l_gameEngine = new GameEngine(l_map);
+        l_gameEngine.getGamePhase().editContinent(l_mapFileName, l_mapFileName); //preload phase command in edit phase
+        String currPhase=l_gameEngine.getGamePhase().getClass().getSimpleName();
+        //assertEquals("Play", currPhase); //Should fail
+        assertNotEquals("Play", currPhase);
+    }
 
+    //checks if game is not skippping phase directly to next phase corretly in postloading
     @Test
-    public void startUpPhaseValidate(){
+    public void inEditPostPhaseValidate(){
         String l_mapFileName = System.getProperty("user.dir") + "/src/main/java/project/app/warzone/Utilities/Maps/" + "europe.map";
         Map l_map = new Map();
         l_map=mapFeatures.readMap(l_mapFileName);
         GameEngine l_gameEngine = new GameEngine(l_map);
-        l_gameEngine.getGamePhase().getClass().getSimpleName();
-        assertEquals("Preload", l_gameEngine.getGamePhase().getClass().getSimpleName());
+        l_gameEngine.getGamePhase().loadMap(l_mapFileName); //postload phase command in edit phase
+        String currPhase=l_gameEngine.getGamePhase().getClass().getSimpleName();
+        //assertEquals("Play", currPhase); //Should fail
+        assertNotEquals("End", currPhase);
     }
-    //checks if game is moving ahead to next phase corretly after preloading
-    @Test
-    public void checkPostloadValidate(){
-        String l_mapFileName = System.getProperty("user.dir") + "/src/main/java/project/app/warzone/Utilities/Maps/" + "europe.map";
-        Map l_map = new Map();
-        l_map=mapFeatures.readMap(l_mapFileName);
-        GameEngine l_gameEngine = new GameEngine(l_map);
-        l_gameEngine.getGamePhase().loadMap(l_mapFileName);
-        l_gameEngine.getGamePhase().next();
-        String nextPhase=l_gameEngine.getGamePhase().getClass().getSimpleName();
-    
-        assertEquals("Postload", nextPhase);
-    }
+
 }
